water_alert/requirements.txt
# Water Alert System — 依賴清單
# pip install -r requirements.txt

# --- 核心 ---
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# --- 串口通訊（LoRa + 雷達）---
pyserial>=3.5

# --- 時序資料庫 ---
influxdb-client>=1.30.0

# --- 圖片處理（雲量辨識）---
opencv-python-headless>=4.8.0
numpy>=1.24.0

# --- HTTP 客戶端 ---
requests>=2.28.0

# --- Pi5 GPIO（僅 Pi 環境需要）---
# lgpio>=0.2.0
# adafruit-circuitpython-dht>=3.7.0

# --- 環境變數 ---
python-dotenv>=1.0.0
